new feature
Enhance debugging for search input field

Added a line to print the "type" attribute of the `txtSrch2` web element to improve debugging information when interacting with the search input field.

Enhance debugging output for search input element

Added a line to print the "type" attribute of the
`txtSrch2` web element in the `SeleniumLearning.cs` file
to improve debugging information.